requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.3
discord.py==2.3.1
frozenlist==1.4.0
idna==3.4
joblib==1.2.0
mock-open==1.4.0
multidict==6.0.4
nltk==3.8.1
numpy==1.24.1
python-dotenv==1.0.0
regex==2022.10.31
requests==2.31.0
simple-term-menu==1.6.1
tqdm==4.64.1
urllib3==2.0.3
yarl==1.9.2
